code reorganisition. renamed action.cpp to enqueue.cpp. consolidate tranpose related actions to action.transpose.cpp
